requirements.txt
django
gunicorn
dj-database-url
psycopg2-binary
whitenoise
pandas
numpy
requests
openai
scikit-learn
xgboost
shap
ta
django-cors-headers
python-dotenv

